Move function from helpers to utils, add unittests
